--- src/app/constant/margin.ts
@@ -1,5 +1,0 @@
-export default {
-  CONTAINER_HORIZONTAL_MARGIN: 16,
-  FLOATING_MARGIN_RIGHT: 16,
-  FLOATING_MARGIN_BOTTOM: 44,
-};
