--- src/app/constant/regex.ts
@@ -1,2 +1,0 @@
-export const CHECK_EMAIL = /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/;
-export const CHECK_JOSHSOFTWARE_EMAIL = /@joshsoftware\.com$/;
