+++ src/app/screens/LoginScreen/index.tsx
@@ -1,15 +1,22 @@
-import React, {useEffect} from 'react';
+import React, {useCallback, useEffect} from 'react';
 import {SafeAreaView, StyleSheet, View} from 'react-native';
 
-// import LoginForm from './LoginForm';
 import Button from '../../components/button';
 import {useLogin} from './login.hooks';
 
-// import colors from '../../constant/colors';
+import {googleSignIn} from '../../services/auth/google.auth';
+
 import {JoshLogo} from '../../constant/icons';
 
 const LoginScreen = () => {
-  const {googleSignInHandler, isLoading} = useLogin();
+  const {mutate, isLoading} = useLogin();
+
+  const googleSignInHandler = useCallback(async () => {
+    const response = await googleSignIn();
+    if (response) {
+      mutate(response);
+    }
+  }, [mutate]);
 
   useEffect(() => {
     googleSignInHandler();
@@ -20,11 +27,6 @@ const LoginScreen = () => {
       <View style={styles.logoContainer}>
         <JoshLogo />
       </View>
-
-      {/* <Text style={styles.loginText}>LOGIN</Text> */}
-      {/* <LoginForm signIn={emailPasswordSignInHandler} isLoading={isLoading} /> */}
-      {/* <Text style={styles.orText}>Or</Text> */}
-
       <View>
         <Button
           type="primary"
@@ -49,19 +51,6 @@ const styles = StyleSheet.create({
     paddingVertical: 90,
     alignItems: 'center',
   },
-  // loginText: {
-  //   fontSize: 22,
-  //   fontWeight: 'bold',
-  //   color: colors.TERTIARY_TEXT,
-  //   paddingVertical: 20,
-  // },
-  // orText: {
-  //   fontSize: 12,
-  //   color: colors.QUATERNARY_TEXT,
-  //   fontWeight: 'bold',
-  //   marginVertical: 14,
-  //   alignSelf: 'center',
-  // },
 });
 
 export default LoginScreen;
