+++ src/app/services/auth/google.auth.ts
@@ -3,8 +3,9 @@ import {
   GoogleSignin,
   statusCodes,
 } from '@react-native-google-signin/google-signin';
-
 import Config from 'react-native-config';
+
+import {INVALID_EMAIL_ERROR} from '../../constant/message';
 
 GoogleSignin.configure({
   webClientId: Config.WEB_CLIENT_ID,
@@ -15,9 +16,18 @@ export const googleSignIn = async () => {
     await GoogleSignin.hasPlayServices();
     const userInfo = await GoogleSignin.signIn();
 
+    if (!userInfo.user.email.endsWith('@joshsoftware.com')) {
+      throw INVALID_EMAIL_ERROR;
+    }
     return userInfo;
   } catch (error: any) {
-    if (error.code === statusCodes.SIGN_IN_CANCELLED) {
+    googleSignOut();
+    if (error === INVALID_EMAIL_ERROR) {
+      Alert.alert(
+        'Login Error',
+        'Only Google accounts from joshsoftware are allowed.',
+      );
+    } else if (error.code === statusCodes.SIGN_IN_CANCELLED) {
       return;
     } else if (error.code === statusCodes.IN_PROGRESS) {
       Alert.alert('', 'sign in is in progress already');
